PLT-1173: Added conditional behavior to metrics to collect in atlas
