+++ common/src/main/java/org/apache/atlas/CommonConfiguration.java
@@ -20,9 +20,11 @@ package org.apache.atlas;
 import io.micrometer.core.aop.TimedAspect;
 import io.micrometer.prometheus.PrometheusConfig;
 import io.micrometer.prometheus.PrometheusMeterRegistry;
+import org.apache.atlas.service.metrics.MetricRegistryCondition;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Conditional;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.EnableAspectJAutoProxy;
 
@@ -47,7 +49,8 @@ public class CommonConfiguration {
     }
 
     @Bean
-    public TimedAspect timedAspect(PrometheusMeterRegistry registry) {
+    @Conditional(MetricRegistryCondition.class)
+    public TimedAspect timedAspect(PrometheusMeterRegistry registry) throws AtlasException {
         return new TimedAspect(registry);
     }
 }
